requirements.txt
absl-py==2.1.0
ale-py==0.11.0
asciitree==0.3.3
asttokens==2.4.1
astunparse==1.6.3
attrs==25.3.0
bcrypt==4.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cvxopt==1.3.2
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
et_xmlfile==2.0.0
executing==2.1.0
Farama-Notifications==0.0.4
fasteners==0.19
fastjsonschema==2.21.1
filelock==3.16.1
fiona==1.10.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==2.4.6
fsspec==2024.10.0
gast==0.6.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
google-pasta==0.2.0
grpcio==1.67.1
gunicorn==23.0.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
h5py==3.12.1
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kagglehub==0.3.12
kaleido==0.2.1
keras==3.6.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multitasking==0.0.11
namex==0.0.8
narwhals==1.35.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numcodecs==0.13.1
numpy==2.1.1
onnxruntime==1.20.1
opencv-python==4.11.0.86
opencv-python-headless==4.8.0.76
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
patsy==1.0.1
peewee==3.18.0
pillow==11.0.0
platformdirs==4.2.2
plotly==6.0.1
prompt_toolkit==3.0.47
protobuf==5.28.3
psutil==6.0.0
psycopg2==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
pygame==2.6.1
Pygments==2.18.0
pyogrio==0.10.0
pyparsing==3.1.4
pyproj==3.7.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
safetensors==0.4.5
sb3_contrib==2.6.0
scikit-image==0.25.2
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
setuptools==75.4.0
shapely==2.0.7
Shimmy==2.0.0
six==1.16.0
soupsieve==2.7
spyder-kernels==3.0.0
stable_baselines3==2.6.0
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.3.30
timm==1.0.15
tokenizers==0.21.0
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.108
ultralytics-thop==2.0.14
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.0
widgetsnbextension==4.0.13
wrapt==1.16.0
yfinance==0.2.57
zarr==2.18.4
zipfile36==0.1.3
